[FEAT] Authentication and Profile

+++ todo_talk.sql
@@ -4,7 +4,7 @@ USE todo_task;
 
 -- Bảng Users: Lưu thông tin người dùng
 CREATE TABLE Users (
-    user_id INT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     username VARCHAR(50) NOT NULL UNIQUE,
     email VARCHAR(100) NOT NULL UNIQUE,
     password_hash VARCHAR(255) NOT NULL,
@@ -18,7 +18,7 @@ CREATE TABLE Users (
 
 -- Bảng Chats: Quản lý các cuộc chat (1-1 hoặc nhóm)
 CREATE TABLE Chats (
-    chat_id INT AUTO_INCREMENT PRIMARY KEY,
+    chat_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     chat_name VARCHAR(100),
     is_group BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -27,9 +27,9 @@ CREATE TABLE Chats (
 
 -- Bảng Chat_Participants: Liên kết user với chat
 CREATE TABLE Chat_Participants (
-    participant_id INT AUTO_INCREMENT PRIMARY KEY,
-    chat_id INT,
-    user_id INT,
+    participant_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    chat_id BIGINT,
+    user_id BIGINT,
     joined_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (chat_id) REFERENCES Chats(chat_id) ON DELETE CASCADE,
     FOREIGN KEY (user_id) REFERENCES Users(user_id) ON DELETE CASCADE,
@@ -38,9 +38,9 @@ CREATE TABLE Chat_Participants (
 
 -- Bảng Messages: Lưu tin nhắn trong chat
 CREATE TABLE Messages (
-    message_id INT AUTO_INCREMENT PRIMARY KEY,
-    chat_id INT,
-    sender_id INT,
+    message_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    chat_id BIGINT,
+    sender_id BIGINT,
     content TEXT NOT NULL,
     is_todo_trigger BOOLEAN DEFAULT FALSE,
     edited_at TIMESTAMP NULL DEFAULT NULL,
@@ -54,9 +54,9 @@ CREATE TABLE Messages (
 
 -- Bảng Message_Reads: Lưu read receipts cho messages
 CREATE TABLE Message_Reads (
-    read_id INT AUTO_INCREMENT PRIMARY KEY,
-    message_id INT,
-    user_id INT,
+    read_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    message_id BIGINT,
+    user_id BIGINT,
     read_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (message_id) REFERENCES Messages(message_id) ON DELETE CASCADE,
     FOREIGN KEY (user_id) REFERENCES Users(user_id) ON DELETE CASCADE,
@@ -65,10 +65,10 @@ CREATE TABLE Message_Reads (
 
 -- Bảng Notifications: Lưu notifications (ví dụ: new message)
 CREATE TABLE Notifications (
-    notification_id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT,
+    notification_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT,
     type ENUM('new_message', 'new_task') NOT NULL,
-    related_id INT NOT NULL, -- message_id hoặc task_id tùy type
+    related_id BIGINT NOT NULL, -- message_id hoặc task_id tùy type
     content VARCHAR(255),
     is_read BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -79,10 +79,10 @@ CREATE TABLE Notifications (
 
 -- Bảng Tasks: Lưu task todo được tạo từ tin nhắn @Todo, gộp proof
 CREATE TABLE Tasks (
-    task_id INT AUTO_INCREMENT PRIMARY KEY,
-    message_id INT,
-    user_id INT,
-    chat_id INT,
+    task_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    message_id BIGINT,
+    user_id BIGINT,
+    chat_id BIGINT,
     description TEXT NOT NULL,
     status ENUM('pending', 'completed') DEFAULT 'pending',
     due_date DATETIME,
